Support listing series for Azure Data Explorer
Renames the source to azure-data-explorer.
