+++ docs/source/azure_data_explorer.asciidoc
@@ -1,0 +1,53 @@
+// SPDX-FileCopyrightText: 2022 Timeseer.AI
+// SPDX-License-Identifier: Apache-2.0
+
+== Azure Data Explorer
+
+Sources with `type = "azure-data-explorer"` configure https://azure.microsoft.com/products/data-explorer/[Azure Data Explorer] sources.
+
+
+```toml
+[source.<name>]
+type = "azure-data-explorer"
+connection_string = "<connection_string>"
+database = "<database>"
+table = "<table>"
+timestamp_column = "ts"
+tag_columns = []
+```
+
+`database`, `table` and `connection_string` are required.
+
+See https://learn.microsoft.com/en-us/azure/data-explorer/kusto/api/connection-strings/kusto[here] to learn more about connection strings.
+
+`timestamp_column` is an optional parameter used to define the name
+of the database column that contains timestamps for the series.
+Defaults to `"ts"`.
+
+Listing time series in an Azure Data Explorer table is supported only when the `tag_columns` parameter is specified.
+Each value in the parameter is a column in the table that will be part of the set of tags of the series.
+Other columns are assumed to be fields.
+
+For example:
+
+```toml
+tag_columns = [
+    "deviceId",
+    "plant",
+    "location",
+]
+```
+
+Metadata in Azure Data Explorer is not supported.
+Use a different `metadata_type` to connect to metadata.
+
+For example:
+
+```toml
+metadata_type = "csv"
+```
+
+The https://github.com/Azure/azure-kusto-python[azure-kusto-data] Python package is required.
+
+The https://github.com/Azure/azure-sdk-for-python/tree/main/sdk/identity/azure-identity#defaultazurecredential[DefaultAzureCredential]
+from the https://pypi.org/project/azure-identity/[azure-identity] package is used for authentication.
