--- docs/source/data_explorer.asciidoc
@@ -1,48 +1,0 @@
-// SPDX-FileCopyrightText: 2022 Timeseer.AI
-// SPDX-License-Identifier: Apache-2.0
-
-== Azure Data Explorer
-
-Sources with `type = "data_explorer"` configure https://azure.microsoft.com/products/data-explorer/[Data Explorer] sources.
-
-
-```toml
-[source.<name>]
-type = "data_explorer"
-connection_string = "<connection_string>"
-database = "<database>"
-table = "<table>"
-timestamps_column = "<timestamps_column>"
-```
-
-`database`, `table` and `connection_string` are required. 
-
-See https://learn.microsoft.com/en-us/azure/data-explorer/kusto/api/connection-strings/kusto[here] to learn more about connection strings.
-
-`timestamp_column` is an optional parameter used to define the name
-of the database column that contains timestamps for the series. Defaults to 'ts'.
-
-Metadata in Data explorer is not supported.
-Use a different `metadata_type` to connect to metadata.
-
-For example:
-
-```toml
-metadata_type = "csv"
-```
-
-=== Connecting to Azure Data Explorer
-
-The https://github.com/Azure/azure-kusto-python[azure-kusto-data] is required.
-The https://github.com/Azure/azure-sdk-for-python/tree/main/sdk/identity/azure-identity#defaultazurecredential[default azure authentication]
-from the https://pypi.org/project/azure-identity/[azure-identity] library is used for authentication.
-
-Example data explorer source definition:
-
-```toml
-[source."Data explorer example"]
-type = "data_explorer"
-connection_string = "https://help.kusto.windows.net/Samples"
-database = "tsai-antewerp"
-table = "sensor-data"
-```
