+++ kukur/source/azure_data_explorer/__init__.py
@@ -1,7 +1,8 @@
 """Connections to azure data explorer data sources from Timeseer."""
-# SPDX-FileCopyrightText: 2021 Timeseer.AI
-#
+
+# SPDX-FileCopyrightText: 2022 Timeseer.AI
 # SPDX-License-Identifier: Apache-2.0
-from .data_explorer import from_config
+
+from .azure_data_explorer import from_config
 
 __all__ = ["from_config"]
